requirements.txt
# Cricbuzz LiveStats - Python Dependencies

# Core Framework
streamlit>=1.28.0

# Database
mysql-connector-python>=8.2.0

# Data Processing
pandas>=2.1.0

# HTTP Requests & API
requests>=2.31.0

# Environment Variables
python-dotenv>=1.0.0
